Move page to user dir
