+++ config/services.yml
@@ -39,8 +39,8 @@ services:
             - '%core.root_path%'
             - '%core.php_ext%'
 
-    ganstaz.gzo.pages:
-        class: ganstaz\gzo\src\pages
+    ganstaz.gzo.page:
+        class: ganstaz\gzo\src\user\page
         arguments:
             - '@cache'
             - '@config'
